Improve change password
Change Password functionality is rewritten in order to make it work and give more and better feedback in case of errors. Following scripts are touched:
- inc_login_scripts.php: a new script, contains some functions from inicio.php because they are now also used by change_password.php. The functions return the current user in case of errrors.
- inicio.php: some functions moved to inc_login_scripts, change the password in case of a password change and return the current user.
- change_password.php: Rewritten: layout and functionality. Original user/password now readonly to prevent modification of other users. Check for correct password, including check for special characters. Disallow action for emergency user and ldap login.
- scripts_dataentry.php: check for special characters, so the User Administration behaves the same as Change Password
