+++ www/htdocs/central/lang/00/admin.tab
@@ -83,6 +83,7 @@ editar=Edit
 editsubc=Literal for subfield substitution
 eliminados=Deleted
 eliminar=Delete
+emergnochgpwd=Emergency login cannot change password with this option
 empty_is_def=Empty selects default folder
 entrar=go
 err_all=You must select ALL or a list o fields
@@ -248,6 +249,7 @@ newoper=New user
 newpass=New password
 no_search_expr_stored=There are not search expressions previously stored
 no_sel_records=You must select the records to be displayed
+nochgpwdldap=Change Password not implemented for LDAP in this script.
 nocopies=No copies selected
 nocopiestoadd=No records found in the copies database to be added to the loanobjects database
 noeditdb=You do not have permission to update this database
@@ -425,10 +427,10 @@ userid=Login
 username=Name
 using_config=Using configured default
 usrmanual=User Manual
-valchars=Only caracters 0-9 and a-z are accepted in login and password
 valdef=Default value
 valdef_save=Save default values
 validextensions=Valid extensions are:
+validpwdchars=Only a-z, A-Z, 0-9 and .,!@#$%^&*?_~\-() are allowed in a password.
 ver=Display
 version=ABCD version
 vistap=Preview
